Update README
Specified a version of the rust compiler that can compile all rust dependencies. There were some updates to some of these dependencies that broke compilation when using not only the rust compiler 1.12.1, but also more recent versions (such as 1.25.0). The cause of the issue seems to be a disparity of versions between sodiumoxide and cryptobox.

Also added a command to instruct the reader how to publish the generated aar to a local maven repository.